[depthMap] SgmParams: Remove cpp file
