+++ src/aliceVision/depthMap/depthMap.cpp
@@ -24,6 +24,55 @@
 namespace aliceVision {
 namespace depthMap {
 
+int computeDownscale(const mvsUtils::MultiViewParams& mp, int scale, int maxWidth, int maxHeight)
+{
+    const int maxImageWidth = mp.getMaxImageWidth() / scale;
+    const int maxImageHeight = mp.getMaxImageHeight() / scale;
+
+    int downscale = 1;
+    int downscaleWidth = mp.getMaxImageWidth() / scale;
+    int downscaleHeight = mp.getMaxImageHeight() / scale;
+
+    while((downscaleWidth > maxWidth) || (downscaleHeight > maxHeight))
+    {
+        downscale++;
+        downscaleWidth = maxImageWidth / downscale;
+        downscaleHeight = maxImageHeight / downscale;
+    }
+
+    return downscale;
+}
+
+void computeScaleStepSgmParams(const mvsUtils::MultiViewParams& mp, SgmParams& sgmParams)
+{
+    const int fileScale = 1; // input images scale (should be one)
+    const int maxSideXY = sgmParams.maxSideXY / mp.getProcessDownscale();
+    const int maxImageW = mp.getMaxImageWidth();
+    const int maxImageH = mp.getMaxImageHeight();
+
+    int maxW = maxSideXY;
+    int maxH = maxSideXY * 0.8;
+
+    if(maxImageW < maxImageH)
+        std::swap(maxW, maxH);
+
+    if(sgmParams.scale == -1)
+    {
+        // compute the number of scales that will be used in the plane sweeping.
+        // the highest scale should have a resolution close to 700x550 (or less).
+        const int scaleTmp = computeDownscale(mp, fileScale, maxW, maxH);
+        sgmParams.scale = std::min(2, scaleTmp);
+    }
+    if(sgmParams.stepXY == -1)
+    {
+        sgmParams.stepXY = computeDownscale(mp, fileScale * sgmParams.scale, maxW, maxH);
+    }
+
+    ALICEVISION_LOG_INFO("Computed SGM parameters:" << std::endl
+                         << "\t- scale: " << sgmParams.scale << std::endl
+                         << "\t- stepXY: " << sgmParams.stepXY);
+}
+
 void getSgmParams(const mvsUtils::MultiViewParams& mp, SgmParams& sgmParams) 
 {
     // get SGM user parameters from MultiViewParams property_tree
